Add checks for a few minor overflow risks
This patch fixes a few locations where specific maliciously crafted
input could cause an arithmetic overflow to bypass a size check. In none
of those instances is the data being parsed actually untrusted for our
use case, so there was no real security risk here. This is just extra
hardening.

BUG=chromium:1441030,b:280378929
TEST=Booted CoachZ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4505019

+++ firmware/2lib/2packed_key.c
@@ -48,7 +48,7 @@ vb2_error_t vb2_unpack_key_buffer(struct vb2_public_key *key,
 
 	/* Validity check key array size */
 	key->arrsize = buf32[0];
-	if (key->arrsize * sizeof(uint32_t) != vb2_rsa_sig_size(key->sig_alg))
+	if ((uint64_t)key->arrsize * sizeof(uint32_t) != vb2_rsa_sig_size(key->sig_alg))
 		return VB2_ERROR_UNPACK_KEY_ARRAY_SIZE;
 
 	key->n0inv = buf32[1];
